bloglist backend 4.22

+++ part4/bloglist/controllers/blogs.js
@@ -1,16 +1,16 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
+// const User = require('../models/user')
 
-const jwt = require('jsonwebtoken')
+// const jwt = require('jsonwebtoken')
 
-const getTokenFrom = request => {
-  const authorization = request.get('authorization')
-  if (authorization && authorization.startsWith('Bearer ')) {
-    return authorization.replace('Bearer ', '')
-  }
-  return null
-}
+// const getTokenFrom = request => {
+//   const authorization = request.get('authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   }
+//   return null
+// }
 
 blogsRouter.get('/', async (_request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -22,13 +22,14 @@ blogsRouter.post('/', async (request, response) => {
 
   const body = request.body
 
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
+  // }
 
-  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
-  if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
-  }
+  // const user = await User.findById(decodedToken.id)
 
-  const user = await User.findById(decodedToken.id)
+  const user = request.user
 
   const blog = new Blog({ ...request.body, author: user.name, user: user._id, likes: body.likes || 0 })
 
@@ -44,22 +45,29 @@ blogsRouter.post('/', async (request, response) => {
 
   response.status(201).json(savedBlog)
 
-  // try {
 
-  // } catch (err) {
-  //   response.status(400).json({ error: 'missing blog data' })
-  // }
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
 
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
-  // try {
-
-  // } catch (error) {
-  //   response.status(410).end()
+  // const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  // if (!decodedToken.id) {
+  //   return response.status(401).json({ error: 'token invalid' })
   // }
+
+  // const user = await User.findById(decodedToken.id)
+
+  const user = request.user
+
+  const blog = await Blog.findById(request.params.id)
+
+  if (blog.user.toString() === user.id) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'Deleting post not possible - user unauthorized' })
+  }
+
 })
 
 blogsRouter.put('/:id', async (request, response) => {
@@ -75,12 +83,7 @@ blogsRouter.put('/:id', async (request, response) => {
 
   const updatedPost = await Blog.findByIdAndUpdate(request.params.id, post, { new: true })
   response.json(updatedPost)
-  // try {
 
-  // }
-  // catch (error) {
-  //   response.status(400).end()
-  // }
 })
